Move messages into separate module tembo_pgmq_python.messages
In order to re-use in async version
