+++ tembo-pgmq-python/tembo_pgmq_python/queue.py
@@ -1,28 +1,9 @@
 from dataclasses import dataclass, field
-from datetime import datetime
 from typing import Optional, List
 from psycopg.types.json import Jsonb
 from psycopg_pool import ConnectionPool
 import os
-
-
-@dataclass
-class Message:
-    msg_id: int
-    read_ct: int
-    enqueued_at: datetime
-    vt: datetime
-    message: dict
-
-
-@dataclass
-class QueueMetrics:
-    queue_name: str
-    queue_length: int
-    newest_msg_age_sec: int
-    oldest_msg_age_sec: int
-    total_messages: int
-    scrape_time: datetime
+from tembo_pgmq_python.messages import Message, QueueMetrics
 
 
 @dataclass
